hostapd: ubus: widen bss_transition_request reassoc_delay to unsigned int
`reassoc_delay` in `hostapd_bss_transition_request()` and
`hostapd_bss_tr_send()` was declared `u8`, but it is read from a
`BLOBMSG_TYPE_INT32` attribute, documented as `int32` in the ubus API
README, and encoded on the wire as a 16-bit little-endian field via
`WPA_PUT_LE16()`.

Two consequences:

- the `reassoc_delay > 65535` guard in `hostapd_bss_tr_send()` can
  never be true with a `u8`, so it is dead code;
- caller-supplied values above 255 are silently truncated by the
  implicit `u32 -> u8` conversion in `blobmsg_get_u32()` -- e.g.
  `reassoc_delay = 1000` is sent as 232, and `reassoc_delay = 256`
  truncates to 0, which drops the `MBO_ATTR_ID_ASSOC_RETRY_DELAY`
  attribute altogether.

Upstream hostapd uses `unsigned int` for the same value, which is why
the `> 65535` check exists there.  Match that type in both the local
declaration and the `hostapd_bss_tr_send()` signature so the existing
range validation is meaningful and values in the documented range are
transmitted verbatim.

+++ package/network/services/hostapd/src/src/ap/ubus.c
@@ -1360,7 +1360,7 @@ static int
 hostapd_bss_tr_send(struct hostapd_data *hapd, u8 *addr, bool disassoc_imminent, bool abridged,
 		    u16 disassoc_timer, u8 validity_period, u8 dialog_token,
 		    struct blob_attr *neighbors, bool include_mbo, u8 mbo_reason,
-		    u8 cell_pref, u8 reassoc_delay)
+		    u8 cell_pref, unsigned int reassoc_delay)
 {
 	struct blob_attr *cur;
 	struct sta_info *sta;
@@ -1506,7 +1506,7 @@ hostapd_bss_transition_request(struct ubus_context *ctx, struct ubus_object *obj
 	bool include_mbo = false;
 	u8 mbo_reason = 0;
 	u8 cell_pref = 0;
-	u8 reassoc_delay = 0;
+	unsigned int reassoc_delay = 0;
 
 	blobmsg_parse(bss_tr_policy, __BSS_TR_DISASSOC_MAX, tb, blob_data(msg), blob_len(msg));
